Move discrete method code to its own file
Matches what we do in the variational case, and we don't use this code much now anyway
